fix: multi-CD discs
I found this when trying to rip CD 2 of a two-disc album.

CDDB identifies individual discs on their own. If there's a two-disc
album, those discs are completely separate with no relation to each
other.

In MusicBrainz, the "release" is the whole album and it can contain
multiple "media" elements to indicate multiple discs.

Since both CDDB and MSDC think in terms of discs, we have to find a
matching release *and* the matching disc and then combine them to one
"disc" in terms that CDDB and MSCD can use.

Way back when I had hardcoded fetching the first "media" element from an
album match. I may not have realised that there's a fundamental
difference between how CDDB and MusicBrainz organise media.